refactor

+++ rust/processor/src/models/multisig_voting_transaction_models/multisig_voting_transaction.rs
@@ -5,10 +5,11 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Clone, Debug, Deserialize, FieldCount, Insertable, Serialize)]
 #[diesel(table_name = multisig_voting_transactions)]
+#[diesel(primary_key(transaction_sequence, wallet_address, owner_address, value))]
 pub struct MultisigVotingTransaction {
     pub wallet_address: String,
     pub owner_address: String,
-    pub transaction_sequence: i64,
+    pub transaction_sequence: i32,
     pub value: bool,
     pub created_at: NaiveDateTime,
 }
